requirements.txt
python-upwork-oauth2
flask
flask_cors
gunicorn
python-dotenv